[backend] Move KojiRepoDescriptor to koji_util
