+++ koschei/backend/repo_util.py
@@ -21,52 +21,7 @@ import hawkey
 import librepo
 import shutil
 
-from functools import total_ordering
-
-from koschei.config import get_config, get_koji_config
-
-
-@total_ordering
-class KojiRepoDescriptor(object):
-    def __init__(self, koji_id, build_tag, repo_id):
-        self.koji_id = koji_id
-        self.build_tag = build_tag
-        self.repo_id = repo_id
-
-    @staticmethod
-    def from_string(name):
-        parts = name.split('-')
-        if len(parts) < 3 or not parts[-1].isdigit():
-            return None
-        return KojiRepoDescriptor(parts[0], '-'.join(parts[1:-1]), int(parts[-1]))
-
-    def __str__(self):
-        return '{}-{}-{}'.format(self.koji_id, self.build_tag, self.repo_id)
-
-    def __hash__(self):
-        return hash((self.koji_id, self.build_tag, self.repo_id))
-
-    def __eq__(self, other):
-        try:
-            return (self.koji_id == other.koji_id and
-                    self.build_tag == other.build_tag and
-                    self.repo_id == other.repo_id)
-        except AttributeError:
-            return False
-
-    def __ne__(self, other):
-        return not self == other
-
-    def __lt__(self, other):
-        return self.repo_id < other.repo_id
-
-    @property
-    def url(self):
-        arch = get_config('dependency.repo_arch')
-        topurl = get_koji_config(self.koji_id, 'topurl')
-        url = '{topurl}/repos/{build_tag}/{repo_id}/{arch}'
-        return url.format(topurl=topurl, build_tag=self.build_tag,
-                          repo_id=self.repo_id, arch=arch)
+from koschei.config import get_config
 
 
 def get_repo(repo_dir, repo_descriptor, download=False):
